A great change regarding trading rule
This system used to trade at close price, and now it trades at open
price, which is more realistic in real world,  by changing the value
updating rule in the core event-driven framework in
trading_backtesting.py. In addtion, a slippage function has been added
to execution_handler model, and it is a normally random slippage.